+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.4.0
+
+- Restore focus to the element that had it, after the modal has closed
+- REVERT `keydown` change in 2.3.0.
+
 ## 2.3.0
 
 - Upgrade dependencies